Move data conversion functions to private module + continue docs
